Updated Go & added spoil command - Added embed to generate repost - Fixed imports from local to module - Made project a module - Other various fixes and changes

+++ manager/commands.go
@@ -17,7 +17,7 @@ type Command struct {
 	Description string
 
 	GuildOnly           bool
-	RequiredPermissions int
+	RequiredPermissions int64
 }
 
 func RegisterCommand(command *Command) {
